requirements.txt
openai>=1.0.0
rich>=13.0.0
requests>=2.31.0
python-dateutil>=2.8.0
